Fix trial notification
